Optimal value double típus lett

+++ PNSDraw/Entities/Solution.cs
@@ -15,7 +15,7 @@ namespace PNSDraw.Entities
 
         public string Title = "";
 
-        public int OptimalValue { get; set; }
+        public double OptimalValue { get; set; }
         public string AlgorithmUsed { get; set; }
 
         public Solution(int index, string title)
